--- apps/client/components/ui/types/testContent.ts
@@ -1,31 +1,0 @@
-export type TitleAndTextProps = {
-  text: {
-    titleText?: string;
-    contentText?: string;
-  };
-};
-
-export type TitleAndMbtiTestProps = {
-  detail: {
-    titleText?: string;
-    imageUrl?: string;
-    squareText?: string;
-  };
-};
-
-export type TitleAndMbtiTestsSmallForSeveralProp = {
-  mbtiTestData: {
-    testCoverDTOList: {
-      commentCount?: number;
-      id?: string;
-      imageUrl?: string;
-      likeCount?: number;
-      playCount?: number;
-      title?: string;
-      type?: string;
-    }[];
-  };
-  style: {
-    titleText: string;
-  };
-};
